Added adjustable delay for slowing down robot movement and redo functionality
Introduced `delayPerStep` to control the speed of robot motion in `Main.m` and `Movement.m`.
Added user input loop in `Main.m` to allow repeating the movement with "redo" command.
Included ability to terminate the loop with "exit" command.
Added robot STEP (also SolidWorks) files and the corresponding XML file to the project.